# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/users/{id}/role": {
      "patch": {
        "operationId": "AdminController_setRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/ban": {
      "post": {
        "operationId": "AdminController_ban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/unban": {
      "post": {
        "operationId": "AdminController_unban",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/service-clients": {
      "post": {
        "operationId": "AdminController_createServiceClient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceClientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/service-clients/{name}": {
      "delete": {
        "operationId": "AdminController_revokeServiceClient",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Botvy",
    "description": "Commands are REST; reads are GraphQL; the live connection is a socket.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "service-token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token"
      }
    },
    "schemas": {
...
      "SetRoleDto": {
        "type": "object",
        "properties": {}
      },
      "BanDto": {
        "type": "object",
        "properties": {}
      },
      "CreateServiceClientDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}